requirements.txt
anyio==3.7.0
archspec @ file:///croot/archspec_1697725767277/work
boltons @ file:///C:/ci_311/boltons_1677729932371/work
Brotli @ file:///C:/ci_311/brotli-split_1676435766766/work
certifi @ file:///C:/b/abs_91u83siphd/croot/certifi_1700501720658/work/certifi
cffi==1.15.1
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
colorama==0.4.6
conda @ file:///C:/b/abs_88drkd35d4/croot/conda_1701719561256/work
conda-content-trust @ file:///C:/b/abs_e3bcpyv7sw/croot/conda-content-trust_1693490654398/work
conda-libmamba-solver @ file:///C:/b/abs_02v6nijijw/croot/conda-libmamba-solver_1700150953092/work/src
conda-package-handling @ file:///C:/b/abs_b9wp3lr1gn/croot/conda-package-handling_1691008700066/work
conda_package_streaming @ file:///C:/b/abs_6c28n38aaj/croot/conda-package-streaming_1690988019210/work
cryptography==41.0.1
DateTime==5.1
distro @ file:///C:/b/abs_a3uni_yez3/croot/distro_1701455052240/work
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.98.0
greenlet==2.0.2
idna @ file:///C:/ci_311/idna_1676424932545/work
iniconfig==2.0.0
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
keyboard==0.13.5
libmambapy @ file:///C:/b/abs_efpsdwt4ya/croot/mamba-split_1698782663578/work/libmambapy
menuinst @ file:///C:/ci_311/menuinst_1678730372782/work
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.1
Pillow==10.1.0
platformdirs @ file:///C:/b/abs_b6z_yqw_ii/croot/platformdirs_1692205479426/work
pluggy==1.2.0
pycosat @ file:///C:/b/abs_31zywn1be3/croot/pycosat_1696537126223/work
pycparser==2.21
PyMySQL==1.1.0
pynput==1.7.6
pyOpenSSL @ file:///C:/b/abs_08f38zyck4/croot/pyopenssl_1690225407403/work
PySocks @ file:///C:/ci_311/pysocks_1676425991111/work
pytest==7.3.2
pytz==2023.3
requests @ file:///C:/b/abs_316c2inijk/croot/requests_1690400295842/work
ruamel.yaml @ file:///C:/ci_311/ruamel.yaml_1676439214109/work
six==1.16.0
sniffio==1.3.0
source==1.2.0
starlette==0.27.0
tabulate==0.9.0
tqdm @ file:///C:/b/abs_f76j9hg7pv/croot/tqdm_1679561871187/work
truststore @ file:///C:/b/abs_55z7b3r045/croot/truststore_1695245455435/work
typing_extensions==4.6.3
urllib3 @ file:///C:/b/abs_9cmlsrm3ys/croot/urllib3_1698257595508/work
win-inet-pton @ file:///C:/ci_311/win_inet_pton_1676425458225/work
zope.interface==6.0
zstandard==0.19.0
